ADDED: API Get Todo List to Generate Excel Report

+++ database/migrations/2025_12_11_070257_create_todos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('title');
             $table->string('assignee')->nullable();
             $table->date('due_date');
-            $table->integer('time_tracked')->default(0);
+            $table->decimal('time_tracked', 10, 2)->default(0);
             $table->enum('status', ['pending', 'open', 'in_progress', 'completed'])->default('pending');
             $table->enum('priority', ['low', 'medium', 'high']);
         });
